'Implemented' but not working :(

+++ src/Security/Security/Secrets/IGf256.cs
@@ -4,7 +4,7 @@
     {
         byte[] Generate(int degree, byte input);
 
-        byte Evaluate(byte[] polynomial, int part);
+        byte Evaluate(byte[] polynomial, byte part);
 
         byte Interpolate(byte[] values);
     }
